requirements.txt
django
biopython
celery
redis
djangorestframework
django-cors-headers
pandas